requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pandas==2.2.2
numpy==1.26.4
groq==0.9.0
python-dotenv==1.0.1
tiktoken==0.8.0
fake-useragent==1.5.1
httpx==0.27.0